Refactor AIChat domain models and services, updating namespaces and class names for consistency
