+++ Backend/Dotnet/Application/AIChat/IAIChatService.cs
@@ -1,19 +1,19 @@
 using Backend.Common.Miscellaneous;
 using Backend.Common.Errors;
 
-namespace Backend.Application.Rag;
+namespace Backend.Application.AIChat;
 
 /// <summary>
-/// Service for RAG (Retrieval-Augmented Generation) query orchestration.
+/// Service for AIChat (Retrieval-Augmented Generation) query orchestration.
 /// Handles prompt template rendering, AI client calls, and result mapping.
 /// </summary>
-public interface IRagService
+public interface IAIChatService
 {
     /// <summary>
-    /// Execute a RAG query asynchronously.
+    /// Execute a AIChat query asynchronously.
     /// </summary>
     /// <param name="model">The domain model containing query, context, and task type.</param>
     /// <param name="cancellationToken">Cancellation token.</param>
-    /// <returns>A result containing either the RAG result or an error.</returns>
-    Task<TryResult<RagResultModel, Error>> QueryAsync(RagDomainModel model, CancellationToken cancellationToken = default);
+    /// <returns>A result containing either the AIChat result or an error.</returns>
+    Task<TryResult<AIChatResultModel, Error>> QueryAsync(AIChatDomainModel model, CancellationToken cancellationToken = default);
 }
